cron/requirements.txt
openai==1.57.1
python-dotenv==1.0.1
supabase==2.10.0
